Create HttpFactory for creating Response objects
