+++ src/Redmine/Client/Psr18Client.php
@@ -14,6 +14,7 @@ use Psr\Http\Message\ServerRequestFactoryInterface;
 use Psr\Http\Message\StreamFactoryInterface;
 use Redmine\Exception\ClientException;
 use Redmine\Http\HttpClient;
+use Redmine\Http\HttpFactory;
 use Redmine\Http\Request;
 use Redmine\Http\Response;
 
@@ -90,37 +91,11 @@ final class Psr18Client implements Client, HttpClient
             $request->getContentType()
         );
 
-        return new class (
+        return HttpFactory::makeResponse(
             $response->getStatusCode(),
             $response->getHeaderLine('Content-Type'),
             strval($response->getBody())
-        ) implements Response {
-            private $statusCode;
-            private $contentType;
-            private $body;
-
-            public function __construct(int $statusCode, string $contentType, string $body)
-            {
-                $this->statusCode = $statusCode;
-                $this->contentType = $contentType;
-                $this->body = $body;
-            }
-
-            public function getStatusCode(): int
-            {
-                return $this->statusCode;
-            }
-
-            public function getContentType(): string
-            {
-                return $this->contentType;
-            }
-
-            public function getContent(): string
-            {
-                return $this->body;
-            }
-        };
+        );
     }
 
     /**
